frontend/requirements.txt
chainlit
requests
supabase
pypdf
python-pptx
python-docx